requirements.txt
Flask==2.3.3
pdfkit==1.0.0
PyPDF2==2.10.8
python-dotenv==1.0.0
reportlab==4.0.4
pdfplumber==0.9.0
